# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Peak Academy API
  version: 1.0.0
servers:
  - url: http://localhost:4000
paths:
...
  /api/auth/me:
    get:
      summary: Current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
...
  /api/payment/initiate:
    post:
      summary: Initiate payment
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payment initiated
...
  /api/payment/history:
    get:
      summary: User payment history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Transactions list
  /api/earnings:
    get:
      summary: Teacher earnings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Earnings list
  /api/earnings/withdrawals:
    post:
      summary: Create withdrawal request
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Withdrawal created
  /api/parent/link-student:
    post:
      summary: Link parent to student
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Link created
  /api/parent/report/{studentId}:
    get:
      summary: Student report
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student report
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT